refactor: Moved methods from ProgramSectionForm to programSectionContent helper
