+++ src/pages/admin/programs/components/program-section-form/ProgramSectionForm.tsx
@@ -4,12 +4,7 @@ import { ImageValues } from '@/types/common/image';
 import { PROGRAMS_TEXT } from '@/const/admin/programs';
 import { renderProgramSection } from '@/utils/functions/render-program-section';
 import styles from './ProgramSectionForm.module.scss';
-import {
-    ProgramSection,
-    ProgramSectionContent,
-    ProgramSectionTemplate,
-    ProgramSectionMode,
-} from '@/types/common/program-sections';
+import { ProgramSection, ProgramSectionTemplate, ProgramSectionMode } from '@/types/common/program-sections';
 import { ContentType } from '@/types/common/programs';
 import {
     getProgramSectionTemplateMaxGroupCount,
@@ -17,6 +12,12 @@ import {
 } from '@/utils/functions/program-section-template-validation/programSectionTemplateValidation';
 import { PROGRAM_SECTION_VALIDATION_FUNCTIONS } from '@/validation/admin/program-schema/program-schema';
 import { getDescriptionAuthorPairsByGroup } from '@/utils/functions/mappers/public/program/get-grouped-program-section-content-pairs';
+import {
+    getContentByType,
+    getDescriptionsInOrder,
+    getFaqPairs,
+    ensureTitleContentAndOnePair,
+} from '@/utils/functions/program-section-content/programSectionContent';
 
 export interface ProgramSectionFormProps {
     section: ProgramSection;
@@ -43,104 +44,6 @@ export interface SectionCancelOptions {
     onAfterDiscard: () => void;
     isTemplateReplacement?: boolean;
 }
-
-const getContentByType = (contents: ProgramSectionContent[], type: ContentType): ProgramSectionContent | undefined => {
-    return contents.find((c) => c.contentType === type);
-};
-
-const getDescriptionsInOrder = (contents: ProgramSectionContent[]) => {
-    return contents.filter((c) => c.contentType === ContentType.Description).sort((a, b) => a.order - b.order);
-};
-
-const getFaqPairs = (contents: ProgramSectionContent[]) => {
-    return contents
-        .filter((c) => c.contentType === ContentType.FaqPair)
-        .sort((a, b) => a.order - b.order)
-        .map((c, index) => ({
-            groupIndex: c.groupIndex ?? index,
-            questionText: c.faqQuestion?.questionText || '',
-            answerText: c.faqQuestion?.answerText || '',
-        }));
-};
-
-const getNextOrder = (contents: ProgramSectionContent[]): number => {
-    return contents.length ? Math.max(...contents.map((c) => c.order)) + 1 : 0;
-};
-
-const ensureTitleContent = (contents: ProgramSectionContent[]): ProgramSectionContent[] => {
-    const hasTitleContent = contents.some((c) => c.contentType === ContentType.Title);
-    if (hasTitleContent) return contents;
-
-    return [
-        ...contents,
-        {
-            contentType: ContentType.Title,
-            order: getNextOrder(contents),
-            title: '',
-        } as any,
-    ];
-};
-
-const ensureDescriptionAuthorPair = (contents: ProgramSectionContent[]): ProgramSectionContent[] => {
-    const pairs = getDescriptionAuthorPairsByGroup(contents);
-    if (pairs.length > 0) return contents;
-
-    const nextOrder = getNextOrder(contents);
-    return [
-        ...contents,
-        {
-            contentType: ContentType.Description,
-            order: nextOrder,
-            groupIndex: 0,
-            description: '',
-        } as any,
-        {
-            contentType: ContentType.Author,
-            order: nextOrder + 1,
-            groupIndex: 0,
-            author: '',
-        } as any,
-    ];
-};
-
-const ensureFaqPair = (contents: ProgramSectionContent[]): ProgramSectionContent[] => {
-    const faqPairs = getFaqPairs(contents);
-    if (faqPairs.length > 0) return contents;
-
-    return [
-        ...contents,
-        {
-            contentType: ContentType.FaqPair,
-            order: getNextOrder(contents),
-            groupIndex: 0,
-            faqQuestion: {
-                questionText: '',
-                answerText: '',
-            },
-        } as any,
-    ];
-};
-
-const PAIRED_TEMPLATES = new Set([
-    ProgramSectionTemplate.SingleTitleDescriptionAuthorPairs,
-    ProgramSectionTemplate.SingleTitleQuestionAnswerPairs,
-]);
-
-const ensureTitleContentAndOnePair = (section: ProgramSection): ProgramSection => {
-    if (!PAIRED_TEMPLATES.has(section.template)) {
-        return section;
-    }
-
-    let contents = ensureTitleContent(section.contents);
-
-    if (section.template === ProgramSectionTemplate.SingleTitleDescriptionAuthorPairs) {
-        contents = ensureDescriptionAuthorPair(contents);
-    } else {
-        contents = ensureFaqPair(contents);
-    }
-
-    return contents === section.contents ? section : { ...section, contents };
-};
 
 export const ProgramSectionForm = ({
     section,
